# swagger.yaml
openapi: 3.0.0
info:
  title: HomeSharing API
  description: REST API для платформы аренды и покупки жилья HomeSharing
  version: "1.0"
...
paths:
...
  /api/me:
    patch:
      tags: [me]
      summary: Обновить роль (isHost)
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isHost]
              properties:
                isHost: { type: boolean }
      responses:
        "200":
          description: Обновлённые флаги пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      isAdmin: { type: boolean }
                      isHost: { type: boolean }
                      isUser: { type: boolean }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  responses:
    Unauthorized:
      description: Не аутентифицирован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...